refactor(httpx): merge fetch.go and jsonapi.go
