+++ internal/engine/httpx/jsonapi.go
@@ -152,3 +152,12 @@ func (c Client) PutJSON(
 	}
 	return c.DoJSON(request, output)
 }
+
+// FetchResource fetches the specified resource and returns it.
+func (c Client) FetchResource(ctx context.Context, URLPath string) ([]byte, error) {
+	request, err := c.NewRequest(ctx, "GET", URLPath, nil, nil)
+	if err != nil {
+		return nil, err
+	}
+	return c.Do(request)
+}
